I have stumbled on what may be a fortunate fork in the road with my work with the npr timeline widget.
While I'm leaving the index.html file alone right now as a sample timeline, I've added the following files:

http://johntynan.com/scripts/timeline/sample.html

sample.html is currently the same as the timeline at index.html, however I've abstracted my javascript into a new single file:

http://johntynan.com/scripts/timeline/nprtimeline.js

Additionally, I've added this proof of concept:

http://johntynan.com/scripts/timeline/generator.html

(based on http://www.mytimelines.net/create-a-timeline/ )


I know for my example timeline the idea was to have a selection list of topics and a date/time range.  That may be a good way to test the ability of the current nprtimeline.js to accept variables.  However, in thinking about the timeline generator page, I figured why re-invent the query generator and simply cut and paste from complete api call into a text box and build the timeline that can be copied and pasted into an html file.